ai-service/requirements.txt
numpy
pandas
scikit-learn
tensorflow
keras
flask
flask-cors
python-dotenv
joblib
requests
scipy
gunicorn
